+++ kms-message/src/kms_message/kms_request.h
@@ -94,7 +94,7 @@ KMS_MSG_EXPORT (char *) kms_request_to_string (kms_request_t *request);
 /* kms_request_to_bytes returns the data for a request.
  * - Returns NULL on error and sets an error on request. */
 KMS_MSG_EXPORT (const uint8_t *)
-kms_request_to_bytes (kms_request_t *request, uint32_t *len);
+kms_request_to_bytes (kms_request_t *request, size_t *len);
 
 #ifdef __cplusplus
 } /* extern "C" */
